Moved autouse fixture away from smoke tests.
